requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
Django==4.2
djangorestframework==3.14.0
joblib==1.2.0
mypy-extensions==1.0.0
nltk==3.8.1
packaging==23.1
pathspec==0.11.1
platformdirs==3.3.0
pytz==2023.3
regex==2023.3.23
sqlparse==0.4.4
tk==0.1.0
tomli==2.0.1
tqdm==4.65.0
